requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==2.0.12
Django==5.1.7
django-admin-interface==0.30.0
django-colorfield==0.13.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
pillow==11.2.1
psycopg2==2.9.10
PyJWT==2.9.0
python-slugify==8.0.4
requests==2.27.1
sqlparse==0.5.3
tabdeal-python==0.4.6
text-unidecode==1.3
tzdata==2025.2
urllib3==1.26.20
websocket-client==1.3.3
